+++ src/components/molecules/mod.rs
@@ -1,5 +1,6 @@
 pub mod action_request_list;
 pub mod actions;
+pub mod create_dao_button;
 pub mod initiative;
 pub mod onboarding;
 pub mod paginator;
@@ -9,6 +10,7 @@ pub use action_request_list::ActionRequestList;
 pub use actions::*;
 pub use initiative::*;
 pub use onboarding::*;
+pub use create_dao_button::CreateDaoButton;
 pub use paginator::Paginator;
 pub use sidebar::Sidebar;
 pub use tabs::Tabs;
